Edits
Removed body list and variable as it had to convey 4 variables. I think I will be using a class of "body" and have the variables assigned to it from there. Then have those variables sent to user's car to clean up the code and make a bit simpler. Plans to have a config file for a rough and dirty "save" "load" functionality. Simulations are done and now need to be implemented in the code, as well as build the rest of the code around them.